Orange: Data Mining Toolbox in Python
“Orange is a machine learning and data mining suite for data analysis
through Python scripting and visual programming. Here we report on the
scripting part, which features interactive data analysis and
component-based assembly of data mining procedures. In the selection
and design of components, we focus on the flexibility of their reuse:
our principal intention is to let the user write simple and clear
scripts in Python, which build upon C++ implementations of
computationally- intensive tasks. Orange is intended both for
experienced users and programmers, as well as for students of data
mining.”