Pre-test and a post-test
The questionnaire consisted of a pre-test and a post-test administered
before the first (Tuesday) and after the last (Friday) activity,
respectively. Both pre- and post-test consist of identical questions,
with the post-test including one additional quantitative question.

The questions are geared to explore conceptions on robotics, the role
of computer science in robotics, and the level of interest to engage in
computer science as a field of study and/or as a hobby. The questions
were designed by the teacher to be age and student appropriate.
Questions were projected to a white-board one-by-one and narrated by
the teacher and responses collected immediately. This approach required
the students to write down answers spontaneously without time for
discussion or deliberation. Questions asked were:

1. What is the purpose of a robot?
2. How can Computer Science “build” robots?
3. How can a robot help humans solve problems?
4. How interested are you to pursue Computer Science in High School and
be-
yond? (1=not at all interested, 5=very interested)
5. How interested are you to purse Computer Science in your spare time,
for ex-
ample reading magazines? (1=not at all interested, 5=very interested)

6. Have you done any research on robotics since last time?
Questions 1-5 were administered for both pre- and post-test. Question 6
was administered only at the post-test.