Learning Analytics and Educational Data Mining in Practice
This paper aims to provide the reader with a comprehensive background
for understanding current knowledge on Learning Analytics (LA) and
Educational Data Mining (EDM) and its impact on adaptive learning. It
constitutes an overview of empirical evidence behind key objectives of
the potential adoption of LA/EDM in generic educational strategic
planning. We examined the literature on experimental case studies
conducted in the domain during the past six years (2008-2013). Search
terms identified 209 mature pieces of research work, but inclusion
criteria limited the key studies to 40. We analyzed the research
questions, methodology and findings of these published papers and
categorized them accordingly. We used non-statistical methods to
evaluate and interpret findings of the collected studies. The results
have highlighted four distinct major directions of the LA/EDM empirical
research. We discuss on the emerged added value of LA/EDM research and
highlight the significance of further implications. Finally, we set our
thoughts on possible uncharted key questions to investigate both from
pedagogical and technical considerations.